requirements.txt
folium
rich
PyQt6
PyQt6-webengine
PyQt6-sip
psycopg2-binary
python-dotenv
geopy
openpyxl
